feat and refactor: add the initial version of the routing
