+++ webapp/src/pages/Root.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Root() {
+    return <p>Proof of concept</p>
+}
