+++ webapp/src/tests/Login.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, fireEvent, screen, waitFor, act } from '@testing-library/react';
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import Login from './Login';
+import Login from '../pages/Login';
 
 const mockAxios = new MockAdapter(axios);
