+++ webapp/src/tests/Register.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, fireEvent, screen, waitFor } from '@testing-library/react';
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import AddUser from './AddUser';
+import AddUser from '../pages/Register';
 
 const mockAxios = new MockAdapter(axios);
